requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0
pytest-html==4.1.1
pytest-bdd==7.0.0
factory-boy==3.3.0
faker==20.1.0
mutmut==2.4.3
locust==2.17.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
pillow==10.1.0
django-filter==23.4
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
coverage==7.3.2
bandit==1.7.5
safety==2.3.4
flake8==6.1.0
black==23.11.0
isort==5.12.0
pre-commit==3.6.0
